% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compoundSymmetry.R
\name{compsym_asr_input}
\alias{compsym_asr_input}
\title{Compound symmetry model for genotype-by-environment interaction using AlphaSimR -
input parameters}
\usage{
compsym_asr_input(
  nEnvs,
  nTraits,
  mean,
  var,
  relMainEffA,
  corA = NULL,
  meanDD = NULL,
  varDD = NULL,
  relMainEffDD = NULL,
  corDD = NULL,
  relAA = NULL,
  relMainEffAA = NULL,
  corAA = NULL
)
}
\arguments{
\item{nEnvs}{Number of environments to be simulated. A minimum of two
environments is required.}

\item{nTraits}{Number of traits to be simulated.}

\item{mean}{A vector of desired mean genetic values for each trait-by-environment
combination (ordered as environments within traits). Simulated traits can
have a different mean for each environment. If the length of \code{mean}
corresponds to \code{nTraits}, however, then traits will be assigned the
same mean for each environment.}

\item{var}{A vector of desired genetic variances for each trait.
Simulated traits are restricted by the compound symmetry model to having
the same variance for each environment (i.e. main effect + GxE interaction)
and the same covariance between each pair of environments (main effect).
\strong{Note:} when \code{useVarA = TRUE} is set in AlphaSimR (default) the
values in \code{var} represent the desired \code{additive} genetic variances,
otherwise they will represent the \code{total} (additive + dominance +
epistatic) genetic variances.}

\item{relMainEffA}{A vector with the desired magnitude of the additive main
effect variance relative to the additive main effect + GxE interaction
variance for each trait. If only one value is provided and \code{nTraits > 1},
then all traits will be assigned the same value. \strong{Note:} \code{0 <
  relMainEffA < 1}.}

\item{corA}{A matrix of additive genetic correlations between more than one
traits. If not defined and \code{nTraits > 1}, a diagonal matrix is assigned.}

\item{meanDD}{A vector of mean dominance degrees for each environment and
trait combination (ordered as environments within traits), similar to
\code{mean}. By default, \code{meanDD = NULL} and dominance is not simulated.}

\item{varDD}{A vector of dominance degree variances for each trait. Simulated
traits have the same dominance degree variance for each environment and the
same dominance degree covariance between each pair of environments, similar
to \code{var}. By default, \code{varDD = NULL}.}

\item{relMainEffDD}{A vector with the desired magnitude of the dominance
degree main effect variance relative to the main effect + GxE interaction
variance for each trait, similar to \code{relMainEffA}. \strong{Note:}
\code{0 < relMainEffDD < 1}. By default, \code{relMaiEffDD = NULL}.}

\item{corDD}{A matrix of dominance degree correlations between more than one
traits, similar to \code{corA}. If not defined and \code{nTraits > 1}, a
diagonal matrix is assigned. By default, \code{corDD = NULL}.}

\item{relAA}{A vector with the magnitude of additive-by-additive (epistatic)
variance relative to additive genetic variance for each trait, that is in a
diploid organism with allele frequency 0.5. Simulated traits have the same
epistatic variance for each environment and same epistatic covariance
between each pair of environments, similar to \code{var}. By default,
\code{relAA = NULL} and epistasis is not simulated.}

\item{relMainEffAA}{A vector with the magnitude of the epistatic main effect
variance relative to the main effect + GxE interaction variance for each
trait, similar to \code{relMainEffA}. \strong{Note:} \code{0 < relMainEffAA < 1}.
By default, \code{relMainEffAA = NULL}.}

\item{corAA}{A matrix of epistatic correlations between  more than one traits,
similar to \code{corA}. If not defined and \code{nTraits > 1}, a diagonal
matrix is assigned. By default, \code{corAA = NULL}.}
}
\value{
A list containing input parameters for AlphaSimR, which is then used
to simulate correlated genetic effects based on a compound symmetry model for
GxE interaction.
}
\description{
Creates a list of simulation parameters for use with AlphaSimR to simulate
genetic values across multiple environments and traits based on a compound
symmetry model for genotype-by-environment (GxE) interaction. By default,
AlphaSimR does not support complex models for GxE interaction. However, its
functionality to simulate correlated genetic values can be utilised for this
purpose by providing the required variance structure. \code{compsym_asr_input}
constructs the required variance structure for a compound symmetry model.
It assumes a separable structure between traits and
environments.
}
\details{
\strong{Note:} for additive traits use AlphaSimR's \code{addTraitA()};
for additive + dominance traits use \code{addTraitAD()}; for additive +
epistatic traits use \code{addTraitAE()}; and for additive + dominance +
epistatic traits use \code{addTraitADE()}. If non-additive effects are to be
simulated, check the \code{useVarA} argument of AlphaSimR's \code{addTrait}
methods.
}
\examples{
# Simulation of genetic values in AlphaSimR for two additive + dominance traits
# and three environments based on a compound symmetry model for GxE interaction.

# 1. Assign genetic architecture of traits
# Mean genetic values and mean dominance degrees for trait 1 in 3 environments
# and trait 2 in 3 environments.
mean <- c(1, 3, 2, 80, 70, 100) # trait 1 by 3 envs, trait 2 by 3 envs.
meanDD <- c(0.1, 0.4) # trait 1 and trait 2, same values assigned to all 3 envs for each trait

# Additive genetic variances (set useVarA=TRUE) and dominance degree variances for traits 1 and 2.
var <- c(0.2, 10)
varDD <- c(0.1, 0.2)

# Relative magnitude of additive and dominance degree main effect variance for traits 1 and 2.
relMainEffA <- c(0.4, 0.6) # different values for traits 1 and 2
relMainEffDD <- 0.8  # same values used for traits 1 and 2

# Additive and dominance degree correlations between traits 1 and 2.
corA <- matrix(c(1.0,  0.3, 0.3,  1.0), ncol = 2)
corDD <- diag(2) # assuming independence between traits

input_asr <- compsym_asr_input(nEnvs = 3, nTraits = 2,
                              mean = mean, var = var,
                              relMainEffA = relMainEffA, corA = corA,
                              meanDD = meanDD, varDD = varDD,
                              relMainEffDD = relMainEffDD, corDD = corDD)


# 2. Use input_asr to simulate genetic values in AlphaSimR based on a compound symmetry model for
#    GxE interaction.

library("AlphaSimR")
FOUNDERPOP <- quickHaplo(nInd = 100,
                         nChr = 6,
                         segSites = 100)

SP <- SimParam$new(FOUNDERPOP)

SP$addTraitAD(nQtlPerChr = 100,
              mean = input_asr$mean,
              var = input_asr$var,
              meanDD = input_asr$meanDD,
              varDD = input_asr$varDD,
              corA = input_asr$corA,
              corDD = input_asr$corDD,
              useVarA = TRUE)             # Variance in var is used as additive variance.
                                          # If FALSE, var = total genetic variance.

pop <- newPop(FOUNDERPOP)


# 3. Create a data frame containing the simulated genetic values for each of the two traits
#    and three environments.

nReps <- c(2, 3, 2)  # Vector with the number of complete replicates in each environment

trial_df <- compsym_asr_output(pop = pop, nEnvs = 3, nReps = nReps,
                               nTraits = 2, effects = TRUE)

}
