% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{error_df_bivar}
\alias{error_df_bivar}
\title{Plot errors - Example data frame}
\format{
A data frame with 700 rows and 6 columns:
\describe{
  \item{env}{Environment number}
  \item{block}{Block number}
  \item{col}{Column number}
  \item{row}{Row number}
  \item{e.Trait1}{Simulated plot errors for Trait 1}
  \item{e.Trait2}{Simulated plot errors for Trait 2}
}
}
\usage{
error_df_bivar
}
\description{
An example data frame with simulated plot errors for two traits in three environments.
Environments 1 and 2 comprise two blocks, while Environment 3 comprises three blocks.
The blocks are aligned in the column direction (side-by-side) and comprise 5 columns and 20 rows.
The data frame was generated using the function \link[FieldSimR]{field_trial_error} with
bivariate interpolation. The simulation is demonstrated in the vignette
\href{https://crwerner.github.io/fieldsimr/articles/spatial_error_demo.html}{Simulation of plot errors and phenotypes in multi-environment field trials}.
}
\keyword{datasets}
