% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldError.R
\name{field_error}
\alias{field_error}
\title{Simulate plot-level error in a multi-environment field trial}
\usage{
field_error(
  nEnvs,
  nTraits,
  nReps,
  nCols,
  nRows,
  plotLength,
  plotWidth,
  repDir = "column",
  varR,
  corR = NULL,
  RcorR = NULL,
  spatialModel = "bivariate",
  propSpatial = 0.5,
  complexity = 12,
  colCor,
  rowCor,
  effects = FALSE
)
}
\arguments{
\item{nEnvs}{Number of environments to be simulated (same as used in
\code{compsym_asr_input} or \code{unstr_asr_output}, where applicable).}

\item{nTraits}{Number of traits to be simulated.}

\item{nReps}{A vector containing the number of complete replicates in each
environment. If only one value is provided,all environments will be assigned
the same value.}

\item{nCols}{A vector containing the number of columns in each environment.
If only one value is provided, all environments will be assigned the same
value.}

\item{nRows}{A vector containing the number of rows in each environment. If
only one value is provided, all environments will be assigned the same value.}

\item{plotLength}{A vector of plot lengths for each environment. If only
one value is provided, all environments will be assigned the same value.}

\item{plotWidth}{A vector of plot widths for each environment. If only one
value is provided, all environments will be assigned the same value.}

\item{repDir}{Character string specifying the direction of replicate blocks.
One of either "column" (side-by-side, the default) or "row" (above-and-below).
Ignored when \code{nReps = 1}.}

\item{varR}{A vector of desired residual variances for each trait-by-environment
combination (ordered as environments within traits). If the length of
\code{varR} corresponds to \code{nTraits}, the traits will be assigned the
same residual variance for each environment.}

\item{corR}{A matrix of residual correlations between more than one traits
with regards to the spatial model only. If not defined and
\code{nTraits > 1}, a diagonal matrix is assigned.}

\item{RcorR}{A matrix of residual correlations between more than one traits
with regards to the random error model only. If not defined and
\code{nTraits > 1}, a diagonal matrix is assigned.}

\item{spatialModel}{Character string specifying the model used to simulate
spatial variation. One of either "Bivariate" (bivariate interpolation, the
default) or "AR1:AR1" (two-dimensional autoregressive process of order one).}

\item{propSpatial}{A vector containing the proportion of residual variance for
the spatial model compared to the total (spatial + random) error variance.
If only one value is provided, all environments will be assigned the same
value. By default, \code{propSpatial = 0.5}.}

\item{complexity}{A single number indicating the complexity of the bivariate
interpolation model. By default, \code{complexity = 12}. Note that much
lower values might result in convergence problems. For more information on
the complexity parameter check \link[akima]{interp}.}

\item{colCor}{A vector of column autocorrelations for each environment used in
the AR1:AR1 model. If only one value is provided, all environments will be
assigned the same value.}

\item{rowCor}{A vector of column autocorrelations for each environment used in
the AR1:AR1 model. If only one value is provided, all environments will be
assigned the same value.}

\item{effects}{When true, a list is returned with additional list entries for
each trait containing the simulated residuals corresponding to the spatial
model and random error within each environment.}
}
\value{
A data-frame containing environment number, block number, column number,
row number and simulated residuals for each trait. When \code{effects = TRUE},
a list is returned with additional list entries for each trait containing
the simulated residuals corresponding to the spatial model and random error
within each environment.
}
\description{
Creates a data frame of residuals for each plot in a simulated multi-environment
field trial testing one or multiple traits. Residuals are based on a spatial
component and a random component. The spatial component can be simulated
using bivariate interpolation based on akima's \link[akima]{interp} function
or based on a separable autoregressive process (AR1:AR1) withinmenvironments.
If multiple traits are simulated, a correlated residual betweenmtraits can be
simulated based assuming a correlation of the spatial residual between traits,
a correlation of the random residual between traits, or ancombination of both.
Between traits and environments, a separable covariance structure is assumed.
}
\examples{
# Simulation of residuals for two traits tested in three environments using
# bivariate interpolation to model spatial variation.

nEnvs <- 3        # number of simulated environments.
nTraits <- 2      # number of simulated traits.

# Field layout
nCols <- 10              # total number of columns in each environment.
nRows <- c(20, 30, 20)   # total number of rows per environment.
plotLength <- 5          # plot length of 5 meters.
plotWidth <- 2           # plot width of 2 meters.
nReps <- c(2, 3, 2)      # number of complete replicates per environment.

# Residual variances for traits 1 and 2
varR <- c(0.4, 15)

# Residual correlations between traits 1 and 2, with regards to spatial model
corR <- matrix(c(1.0,  0.2, 0.2, 1.0), ncol = 2)

plot_df <- field_error(nEnvs = nEnvs, nTraits = nTraits,
                      nCols = nCols, nRows = nRows, plotLength = plotLength,
                      plotWidth = plotWidth, nReps = nReps, repDir = "row",
                      varR = varR, corR = corR, spatialModel = "bivariate",
                      propSpatial = 0.6, complexity = 14, effects = TRUE)

}
