% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapField.R
\name{map_field}
\alias{map_field}
\title{Map field trial}
\usage{
map_field(df, env, trait, borders = TRUE)
}
\arguments{
\item{df}{Data frame containing at least columns "env", "row", "col" and
the trait to be plotted. If df contains a column "block", the blocks of
full replicates are also indicated, if \code{borders = TRUE}. If df is a
list, the first element of the list will be used by default. To use a
different list element, this has to be specified.}

\item{env}{Environment id of the field to be plotted.}

\item{trait}{Column identifier of the trait to be plotted.}

\item{borders}{When true, blocks of full replicates are indicated.}
}
\value{
A field plot of the input trait represented by a colour gradient from
red (low value) to green (high value).
}
\description{
Plots the values of some input trait in the field represented by a colour
gradient going from red (low value) to green (high value). The input  trait
must have a unique value for each row x column combination within an
environment. The function \code{map_field} was created to take data frames
generated
with \link[FieldSimR]{field_error} as an input, but can work with every data
frame that contains at least the columns "env", "col", "row" and the trait to
be plotted (e.g., the simulated plot-level residual "e.Trait.1").  If the
input data frame contains a column "blocks", the blocks of full replicated
will be indicates in the field plot.
}
\examples{
# Simulation of residuals for two traits tested in three environments using
# bivariate interpolation to model spatial variation.

nEnvs <- 3 # number of simulated environments.
nTraits <- 2 # number of simulated traits.

# Field layout
nCols <- 10 # total number of columns in each environment.
nRows <- c(20, 30, 20) # total number of rows per environment.
plotLength <- 5 # plot length of 5 meters.
plotWidth <- 2 # plot width of 2 meters.
nReps <- c(2, 3, 2) # number of complete replicates per environment.

# Residual variances for traits 1 and 2
varR <- c(0.4, 15)

# Residual correlations between traits 1 and 2, with regards to spatial model
corR <- matrix(c(1.0, 0.2, 0.2, 1.0), ncol = 2)

plot_df <- field_error(
  nEnvs = nEnvs, nTraits = nTraits,
  nCols = nCols, nRows = nRows, plotLength = plotLength,
  plotWidth = plotWidth, nReps = nReps, repDir = "row",
  varR = varR, corR = corR, spatialModel = "bivariate",
  propSpatial = 0.6, complexity = 14, effects = FALSE
)

# Plot the simulated error for trait 2 in environment 2.
map_field(plot_df, env = 2, trait = "e.Trait.2")

}
