% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_variances.R
\name{measure_variances}
\alias{measure_variances}
\title{Measure variances from a covariance matrix}
\usage{
measure_variances(mat, estimate = FALSE)
}
\arguments{
\item{mat}{A symmetric \code{n x n} positive (semi)-definite variance matrix.}

\item{estimate}{When \code{TRUE} (default is \code{FALSE}), the variance components are treated
as estimates and sample corrections are applied.}
}
\value{
A table which partitions the total variance into main effect and interaction variance,
  heterogeneity of variance and lack of correlation, and non-crossover and crossover interaction.
  When \code{correction = TRUE}, a sample correction is applied to the variance components.
}
\description{
Creates a table of variance components derived from a user-defined covariance matrix.
}
\examples{
# Generate a structured covariance matrix and then measure the partitioning of variance.

diag_mat <- rand_diag_mat(
  n = 10
)
cor_mat <- struc_cor_mat(
  n = 10,
)

cov_mat <- sqrt(diag_mat) \%*\% cor_mat \%*\% sqrt(diag_mat)

measure_variances(
  mat = cov_mat
)

}
