% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_trial_effects}
\alias{plot_trial_effects}
\title{Plot field trial effects}
\usage{
plot_trial_effects(df, env, effect, borders = TRUE)
}
\arguments{
\item{df}{A data frame containing at least the columns "env", "row", "col", and the effect to
be plotted. If \code{df} contains a a "block" column, block borders will be indicated if
\code{borders = TRUE}. If \code{df} is a list, the first list entry will be used by default.
To use a different list entry, the entry to be used needs to be specified.}

\item{env}{The ID of the environment to be plotted.}

\item{effect}{The column ID of the effect to be plotted.}

\item{borders}{When true (default), block boarders are indicated.}
}
\value{
A field plot of the input effect represented by a colour gradient from
  red (low value) to green (high value).
}
\description{
Plots trial effects (e.g., genetic values, phenotypic values, errors) at the plot-level
represented by a colour gradient ranging from red (low value) to green (high value). The effect
to be plotted must have a unique value for each row x column combination. \cr
\code{plot_trial_effects} directly takes data frames generated with
\link[FieldSimR]{field_trial_error} as an input, but will also work with other data frames that
contain at least the columns "env", "col", "row", and the effect to be plotted. If the data
frame contains a "block" column, block borders will be indicated if \code{borders = TRUE}.
}
\examples{
# Simulation of plot-level errors for two traits tested in three environments using bivariate
# interpolation to model spatial variation.

n_envs <- 3 # Number of simulated environments.
n_traits <- 2 # Number of simulated traits.

# Field layout
n_cols <- 10 # Total number of columns in each environment.
n_rows <- c(20, 30, 20) # Total number of rows in each environment.
plot_length <- 5 # Plot length set to 5 meters in each environment.
plot_width <- 2 # Plot width set to 2 meters in each environment.
n_reps <- c(2, 3, 2) # Number of complete replicates (blocks) per environment.

# Error variances for traits 1 and 2.
var_R <- c(0.4, 15)

# Spatial error correlations between traits 1 and 2.
cor_R <- matrix(c(
  1.0, 0.2,
  0.2, 1.0
),
ncol = 2
)

error_df <- field_trial_error(
  n_envs = n_envs,
  n_traits = n_traits,
  n_cols = n_cols,
  n_rows = n_rows,
  plot_length = plot_length,
  plot_width = plot_width,
  n_reps = n_reps,
  rep_dir = "row",
  var_R = var_R,
  cor_R = cor_R,
  spatial_model = "bivariate",
  prop_spatial = 0.6,
  complexity = 14,
  return_effects = TRUE
)

# Plot the simulated error for trait 2 in environment 2.
plot_trial_effects(error_df,
  env = 2,
  effect = "e.Trait.2"
)
}
