% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{qq_plot}
\alias{qq_plot}
\title{Q-Q plot}
\usage{
qq_plot(df, effect, labels = FALSE)
}
\arguments{
\item{df}{A data frame containing the effect to be compared to a normal distribution using a
Q-Q plot.}

\item{effect}{The name of the effect to be compared to a normal distribution using a Q-Q plot.}

\item{labels}{When FALSE (default), data points without labels are plotted. When TRUE, column
and row labels are shown in the Q-Q plot. This requires additional columns "col" and "row" in
the data frame.}
}
\value{
A Q-Q plot with the x- and y-axes displaying the theoretical quantile and the sample
  quantile of the effect to be used, respectively.
}
\description{
Creates a Q-Q plot (quantile-quantile plot) for graphical comparison of a normal theoretical
quantile and the quantile of the effect to be used.
}
\examples{
# Q-Q plot of the simulated total error for trait 2 in environment 2 provided in the example
# data frame 'df_error_bivar'.

error_df <- df_error_bivar[df_error_bivar$env == 2, ]

qq <- qq_plot(
  df = error_df,
  effect = "e.Trait.2",
  labels = TRUE
)

# Q-Q plot
qq

# Extraction of a data frame containing the theoretical and sample quantiles of
# the effect.
qq_df <- qq$data

}
