% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_error_bivar}
\alias{df_error_bivar}
\title{Plot errors - example data frame}
\format{
A data frame with 800 rows of data and the following 6 columns:
\describe{
  \item{env}{Environment number}
  \item{block}{Block number}
  \item{col}{Column number}
  \item{row}{Row number}
  \item{e.Trait.1}{Simulated plot error for trait 1}
  \item{e.Trait.2}{Simulated plot error for trait 2}
}
}
\usage{
df_error_bivar
}
\description{
An example data frame of simulated plot errors for two traits
across three environments. The data frame was generated using the function
\link[FieldSimR]{field_trial_error} with bivariate interpolation.
}
\keyword{datasets}
