% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_error_bivar}
\alias{df_error_bivar}
\title{Field trial error example data frame}
\format{
A data frame with 700 rows and 5 columns:
\describe{
  \item{env}{Environment id}
  \item{block}{Block id}
  \item{col}{Column id}
  \item{row}{Row id}
  \item{e.Trait.1}{Simulated plot-level error for trait 1}
  \item{e.Trait.2}{Simulated plot_level error for trait 2}
}
}
\usage{
df_error_bivar
}
\description{
An example data frame of simulated field trial error values at the plot-level for two traits
tested in three environments. The data frame was generated using the function
\link[FieldSimR]{field_trial_error} with bivariate interpolation.
}
\keyword{datasets}
