% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_error_bivar}
\alias{df_error_bivar}
\title{Plot errors - example data frame}
\format{
A data frame with 800 rows and 6 columns:
\describe{
  \item{env}{Environment number}
  \item{block}{Block (replicate) number}
  \item{col}{Column number}
  \item{row}{Row number}
  \item{e.Trait.1}{Simulated plot error for trait 1}
  \item{e.Trait.2}{Simulated plot error for trait 2}
}
}
\usage{
df_error_bivar
}
\description{
An example data frame of simulated plot errors for two traits across three locations. The first
two locations include three replicates (blocks), and the third location includes two replicates.
Each replicate comprises 10 rows and 10 columns. The data frame was generated using the
function \link[FieldSimR]{field_trial_error} with bivariate interpolation. The simulation of
the plot errors is shown in the vignette on the
\href{https://crwerner.github.io/fieldsimr/articles/spatial_error_demo.html}{Simulation of plot errors and phenotypes for a multi-environment plant breeding trial}.
}
\keyword{datasets}
