% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_trial_error.R
\name{field_trial_error}
\alias{field_trial_error}
\title{Simulate plot errors in plant breeding trials}
\usage{
field_trial_error(
  n_envs = 1,
  n_traits = 1,
  n_reps = 2,
  n_cols = 10,
  n_rows = 20,
  rep_dir = "col",
  var_R = 1,
  S_cor_R = NULL,
  R_cor_R = NULL,
  E_cor_R = NULL,
  spatial_model = "Bivariate",
  complexity = NULL,
  plot_length = 10,
  plot_width = 2,
  col_cor = 0.4,
  row_cor = 0.6,
  prop_spatial = 0.5,
  prop_ext = 0,
  ext_dir = "row",
  ext_col_cor = 0,
  ext_row_cor = 0,
  return_effects = FALSE
)
}
\arguments{
\item{n_envs}{Number of environments to be simulated (same as for \code{compsym_asr_input}
or \code{unstr_asr_output}, where applicable). By default \code{n_envs = 1}.}

\item{n_traits}{Number of traits to be simulated. By default, \code{n_traits = 1}.}

\item{n_reps}{A vector specifying the number of complete replicates in each environment. If
only one value is provided, all environments will be assigned the same number of replicates.
By default, \code{n_reps = 2}.}

\item{n_cols}{A vector specifying the total number of columns in each environment. If only one
value is provided, all environments will be assigned the same number of columns. By default,
\code{n_cols = 10}.}

\item{n_rows}{A vector specifying the total number of rows in each environment. If only one
value is provided, all environments will be assigned the same number of rows. By default
\code{n_rows = 20}.}

\item{rep_dir}{A vector specifying the direction of replicate blocks in each environment. Use
"col" for a side-by-side arrangement (default), "row" for an above-and-below arrangement, or
NA if only one replicate block is simulated. If only one value is provided, all environments
will be assigned the same replicate block direction (where applicable).}

\item{var_R}{A vector of error variances for each environment within trait combination. If only
one value is provided, all environment within trait combinations will be assigned the same
error variance.}

\item{S_cor_R}{A matrix of spatial error correlations between more than one trait. If not
specified, a diagonal matrix is constructed.}

\item{R_cor_R}{A matrix of random error correlations between more than one trait. If not
specified, a diagonal matrix is constructed.}

\item{E_cor_R}{A matrix of extraneous error correlations between more than one trait. If not
specified, a diagonal matrix is constructed. Note that the same correlation between traits
is used for the column and row errors (where applicable).}

\item{spatial_model}{A character string specifying the model used to simulate the two-dimensional
spatial error term. One of either "Bivariate" (bivariate interpolation, the default) or "AR1:AR1"
(separable first-order autoregressive process).}

\item{complexity}{A vector specifying the complexity of the bivariate interpolation in each
environment. If only one value is provided, all environments will be assigned the same complexity.
If not specified and \code{spatial_model = "Bivariate"}, the complexity is set to the maximum
number of columns and rows in each environment. This generally provides good results. See
\link[interp]{interp} for further details.}

\item{plot_length}{A vector of plot lengths (column direction, usually longer side) for each
environment. If only one value is provided, all environments will be assigned the same plot
length. Only required when \code{spatial_model = "Bivariate"}. By default, \code{plot_length = 10}.}

\item{plot_width}{A vector of plot widths (row direction, usually shorter side) for each environment.
If only one value is provided, all environments will be assigned the same plot width. Only
required when \code{spatial_model = "Bivariate"}. By default, \code{plot_width = 2}.}

\item{col_cor}{A vector of column autocorrelations for each environment used in the AR1:AR1
spatial error model. If only one value is provided, all environments will be assigned the same
column autocorrelation. Only required when \code{spatial_model = "AR1:AR1"}. By default,
\code{col_cor = 0.4}.}

\item{row_cor}{A vector of row autocorrelations for each environment used in the AR1:AR1 spatial
error model. If only one value is provided, all environments will be assigned the same row
autocorrelation. Only required when \code{spatial_model = "AR1:AR1"}. By default, \code{row_cor = 0.6}.}

\item{prop_spatial}{A vector specifying the proportion of spatial error variance to total error
variance (spatial + random + extraneous) for each environment within trait combination. If only
one value is provided, all environment within trait combinations will be assigned the proportion
of sptial error variance. By default, \code{prop_spatial = 0.5}.}

\item{prop_ext}{A vector specifying the proportion of extraneous error variance to total error
variance (spatial + random + extraneous) for each environment within trait combination. If only
one value is provided, all environment within trait combinations will be assigned the proportion
of extraneous error variance. By default, \code{prop_ext = 0}.}

\item{ext_dir}{A vector specifying the direction of extraneous variation for each environment.
Use "col" to simulate variation in the column direction, "row" (default) for variation in the
row direction, "both" for variation in both directions, or NA if zero extraneous variation is
simulated. When \code{ext_dir = "both"}, half the variance is assigned to the columns and
half is assigned to the rows. If only one value is provided, all environments will be
assigned the same direction (where applicable).}

\item{ext_col_cor}{A vector of column autocorrelations for each environment used in the extraneous
error model. If only one value is provided, all environments will be assigned the same column
autocorrelation. By default, \code{ext_col_cor = 0}.}

\item{ext_row_cor}{A vector of row autocorrelations for each environment used in the extraneous
error model. If only one value is provided, all environments will be assigned the same row
autocorrelation. By default, \code{ext_row_cor = 0}.}

\item{return_effects}{When TRUE, a list is returned with additional entries for each trait
containing the spatial, random and extraneous errors. By default, return_effects = FALSE.}
}
\value{
A data frame with columns "env", "block", "col" and "row", as well as the
  simulated error for each trait. When \code{return_effects = TRUE}, a list is returned with
  additional columns for each trait providing the spatial, random and extraneous errors.
}
\description{
Creates a data frame with simulated plot errors for one or more traits in plant breeding
trials across one or more environments. The plot errors consist of spatial error, random
error and extraneous error. The spatial error is simulated according to either 1) bivariate
interpolation using the \link[interp]{interp} function of the package 'interp', or 2) a
separable first-order autoregressive process (AR1:AR1). The random error is simulated using
an independent process. The extraneous error is simulated as the sum of column and/or row terms,
where the user can choose from independent or correlated processes. The spatial, random and
extraneous errors are combined according to a user-defined ratio.
\cr
For multiple traits, correlated errors can be simulated assuming 1) correlated spatial
error between traits, 2) correlated random error between traits, 3) correlated extraneous
error between traits, or 4) some combination of 1-3. \cr
A separable correlation structure is assumed between traits and environments, but different
variances can be assigned to different environment within trait combinations.
}
\examples{
# Simulation of plot errors for two traits in three environments using a bivariate
# interpolation model for spatial variation.

n_envs <- 3
n_traits <- 2

# Field layout
n_cols <- 10
n_rows <- c(20, 30, 30)
rep_dir <- c("col", "row", "row")
plot_length <- 10
plot_width <- 2
n_reps <- c(2, 3, 3)

# Error variances for environment within trait combinations.
var_R <- c(0.2, 0.4, 0.6, 10, 15, 20)

# Spatial error correlations between traits 1 and 2.
S_cor_R <- matrix(
  c(
    1.0, 0.2,
    0.2, 1.0
  ),
  ncol = 2
)

# Structure of simulated error.
prop_spatial <- 0.4
prop_ext <- 0.2
ext_dir <- "row"
ext_row_cor <- -0.6

error_df <- field_trial_error(
  n_envs = n_envs,
  n_traits = n_traits,
  n_reps = n_reps,
  n_cols = n_cols,
  n_rows = n_rows,
  rep_dir = "row",
  var_R = var_R,
  S_cor_R = S_cor_R,
  spatial_model = "Bivariate",
  plot_length = plot_length,
  plot_width = plot_width,
  prop_spatial = prop_spatial,
  prop_ext = prop_ext,
  ext_dir = ext_dir,
  ext_row_cor = ext_row_cor,
  return_effects = TRUE
)
}
