% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_effects}
\alias{plot_effects}
\title{Graphics for plot-level effects}
\usage{
plot_effects(df, env, effect, blocks = TRUE)
}
\arguments{
\item{df}{A data frame containing the columns "env", "row", "col", and the effect to be plotted.
If \code{df} contains a column named "block", then block borders will distinguish the blocks
if \code{blocks = TRUE}. If \code{df} is a list, only the first entry will be used unless
otherwise specified.}

\item{env}{The name of the environment to be plotted.}

\item{effect}{The name of the effect to be plotted.}

\item{blocks}{When TRUE (default), blocks are distinguished with block borders.}
}
\value{
Graphic of the field array, where the colour gradient ranges from red (low value) to
  green (high value) of the effect
}
\description{
Graphically displays plot-level effects (e.g., phenotypic values, genetic values, errors) onto
a field array, where the colour gradient ranges from red (low value) to green (high value). \cr
This function requires a data frame generated with \link[FieldSimR]{field_trial_error} as an
input, or any data frame with columns named "env", "col", "row", and the effect to be displayed.
If the data frame contains a column named "block", then block borders will distinguish the
blocks if \code{blocks = TRUE}.
}
\examples{
# Plot the simulated error for trait 2 in environment 2 of the field trial error example data
# frame 'df_error_bivar'.

error_df <- df_error_bivar

plot_effects(
  error_df,
  env = 2,
  effect = "e.Trait.2"
)
}
