% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Graphics for matrices}
\usage{
plot_matrix(mat, order = FALSE, labels = FALSE)
}
\arguments{
\item{mat}{A symmetric matrix.}

\item{order}{When \code{TRUE} (default is \code{FALSE}), the function \code{agnes} of the R package
\href{https://cran.r-project.org/package=cluster}{cluster} is used with default arguments to
order the matrix based on a dendrogram.}

\item{labels}{When \code{TRUE} (default is \code{FALSE}), variable labels are displayed.}
}
\value{
A graphical field array with x- and y-axes displaying the variable numbers,
 and colour gradient ranging from blue (low value) to red (high value).
}
\description{
Creates a heatmap for a symmetric matrix (e.g., correlation or covariance matrix).
}
\examples{
# Display a random correlation matrix.

corA <- rand_cor_mat(
  n = 10,
  min.cor = -1,
  max.cor = 1
)

plot_matrix(
  mat = corA,
  order = TRUE,
  labels = TRUE
)

}
