% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_trial_error.R
\name{plot_trial_effects}
\alias{plot_trial_effects}
\title{Plot field trial effects}
\usage{
plot_trial_effects(df, env, trait, borders = TRUE)
}
\arguments{
\item{df}{Data frame containing at least columns "env", "row", "col" and
the trait to be plotted. If df contains a column "block", the blocks of
full replicates are also indicated, if \code{borders = TRUE}. If df is a
list, the first element of the list will be used by default. To use a
different list element, this has to be specified.}

\item{env}{Environment id of the field to be plotted.}

\item{trait}{Column identifier of the trait to be plotted.}

\item{borders}{When true, blocks of full replicates are indicated.}
}
\value{
A field plot of the input trait represented by a colour gradient from
  red (low value) to green (high value).
}
\description{
Plots the values of some input trait in the field represented by a colour
gradient going from red (low value) to green (high value). The input  trait
must have a unique value for each row x column combination within an
environment. The function \code{plot_trial_effects} was created to take data frames
generated
with \link[FieldSimR]{field_trial_error} as an input, but can work with every data
frame that contains at least the columns "env", "col", "row" and the trait to
be plotted (e.g., the simulated plot-level residual "e.Trait.1").  If the
input data frame contains a column "blocks", the blocks of full replicated
will be indicates in the field plot.
}
\examples{
# Simulation of residuals for two traits tested in three environments using bivariate
# interpolation to model spatial variation.

n_envs <- 3 # number of simulated environments.
n_traits <- 2 # number of simulated traits.

# Field layout
n_cols <- 10 # total number of columns in each environment.
n_rows <- c(20, 30, 20) # total number of rows in each environment.
plot_length <- 5 # plot length set to 5 meters.
plot_width <- 2 # plot width set to 2 meters.
n_reps <- c(2, 3, 2) # number of complete replicates (blocks) per environment.

# Residual variances for traits 1 and 2.
var_R <- c(0.4, 15)

# Residual correlations between traits 1 and 2, with regards to spatial model.
cor_R <- matrix(c(1.0, 0.2,
                  0.2, 1.0),
                  ncol = 2)

error_df <- field_trial_error(n_envs = n_envs,
                              n_traits = n_traits,
                              n_cols = n_cols,
                              n_rows = n_rows,
                              plot_length = plot_length,
                              plot_width = plot_width,
                              n_reps = n_reps,
                              rep_dir = "row",
                              var_R = var_R,
                              cor_R = cor_R,
                              spatial_model = "bivariate",
                              prop_spatial = 0.6,
                              complexity = 14,
                              effects = TRUE)

# Plot the simulated error for trait 2 in environment 2.
plot_trial_effects(error_df,
                   env = 2,
                   trait = "e.Trait.2")

}
