% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{qq_plot}
\alias{qq_plot}
\title{Construct a qqplot}
\usage{
qq_plot(df, effect, labels = TRUE, plot = TRUE)
}
\arguments{
\item{df}{A data frame containing the columns "col", "row", and the effect to be plotted.}

\item{effect}{The name of the effect to be plotted.}

\item{labels}{When TRUE (default), the column and row labels are inserted onto the qqplot.
Otherwise, data points without labels are plotted. In this case, the data frame does not
require the columns "col" and "row", just the effect to be plotted.}

\item{plot}{When TRUE (default), the qqplot is displayed graphically.
Otherwise, a data frame is returned.}
}
\value{
Graphic of the qqplot, where the x- and y- axes display the theoretical and
  sample quantiles. When \code{plot = FALSE}, a data frame is returned with the column and row
  displacements as well as the estimated semi-variances.
}
\description{
Construct a qqplot
}
\examples{
# Plot the qqplot for the spatial error component in the data frame error_df.

qq_plot(
  error_df, # what is error_df ?
  effect = "e_spat",
  labels = TRUE,
  plot = TRUE,
)
}
