% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapField.R
\name{rand_cor_mat}
\alias{rand_cor_mat}
\title{Simulate a correlation matrix between traits or environments}
\usage{
rand_cor_mat(p, minCor = -1, maxCor = 1, digits = 2)
}
\arguments{
\item{p}{A scalar defining the dimension of the correlation matrix}

\item{minCor}{A scalar defining the minimum correlation. By default,
minCor = -1.}

\item{maxCor}{A scalar defining the maximum correlation. By default,
maxCor = 1.}

\item{digits}{Decimal digits. By default, digits = 2.}
}
\value{
A p x p correlation matrix.
}
\description{
Creates a \code{p x p} correlation matrix, which can be used for traits or
environments between genotypes or residuals. An upper and lower bound of the
sampled correlations can be set.
}
\examples{
# corA <- rand_cor_mat(10, minCor = -0.2, maxCor = 0.8)

}
