% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{sample_variogram}
\alias{sample_variogram}
\title{Construct a sample variogram}
\usage{
sample_variogram(df, effect, plot = TRUE, min_np = 30)
}
\arguments{
\item{df}{A data frame containing the columns "col", "row", and the effect to be plotted.}

\item{effect}{The name of the effect to be plotted.}

\item{plot}{When TRUE (default), the sample variogram is displayed graphically.
Otherwise, a data frame is returned.}

\item{min_np}{The minimum number of pairs that semi-variances will be displayed for.
By default, \code{min_np = 30}.}
}
\value{
Graphic of the sample variogram, where the x- and y- axes display the row and
  column displacements and the z-axis displays the semi-variance (variogram ordinates).
  When \code{plot = FALSE}, a data frame is returned with the column and row displacements
  as well as the estimated semi-variances.
}
\description{
Construct a sample variogram
}
\examples{
# Plot the sample variogram for the spatial error component in the data frame error_df.

sample_variogram(
  error_df,
  effect = "e_spat",
  plot = TRUE,
)
}
