% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{theoretical_variogram}
\alias{theoretical_variogram}
\title{Construct a theoretical variogram}
\usage{
theoretical_variogram(
  n_cols,
  n_rows,
  var_R = 1,
  prop_spatial = 0.5,
  col_cor,
  row_cor,
  plot = TRUE
)
}
\arguments{
\item{n_cols}{The total number of columns.}

\item{n_rows}{The total number of rows.}

\item{var_R}{The total error variance. By default, \code{var_R = 1}.}

\item{prop_spatial}{The proportion of spatial error variance to total error
variance (spatial + random). By default, \code{prop_spatial = 0.5}.}

\item{col_cor}{The column autocorrelation value.}

\item{row_cor}{The row autocorrelation value.}

\item{plot}{When TRUE (default), the theoretical variogram is displayed graphically.
Otherwise, a data frame is returned.}
}
\value{
Graphic of the theoretical variogram, where the x- and y- axes display the row and
  column displacements and the z-axis displays the semi-variance (variogram ordinates).
  When \code{plot = FALSE}, a data frame is returned with the column and row displacements
  as well as the theoretical semi-variances.
}
\description{
Construct a theoretical variogram
}
\examples{
# Plot a theoretical variogram for a field with 10 columns and 20 rows,
# using column and row autocorrelations of 0.4 and 0.8.

theoretical_variogram(
  n_cols = 10,
  n_rows = 20,
  var_R = 1,
  prop_spatial = 0.5,
  col_cor = 0.4,
  row_cor = 0.8,
  plot = TRUE
)
}
